fix: manual latex build with proper build dir

+++ src/_tex-to-pdf/build_manually.sh
@@ -1,15 +1,21 @@
 #!/bin/sh
 
+BUILD_DIR=$1
+
+if [ $# != 1 ] || [ ! -d "$BUILD_DIR" ]; then
+	echo "Usage: ${0##*/} path/to/build-dir/"
+	exit 1
+fi
+
 cd "${0%/*}" || exit 1
 rm -rf out
 mkdir -p out
 
-build_dir=$(lektor project-info --output-path)
-echo "\\\\def\\\\buildDir{$build_dir}" > out/builddir.tex
+echo "\\\\def\\\\buildDir{$BUILD_DIR}" > out/builddir.tex
 
-find "$build_dir" -name "*.tex" | while read -r x; do
+find "$BUILD_DIR" -name "*.tex" | while read -r x; do
 	bname=$(basename "${x%.tex}") # remove extension and parent dir
-	shortname=${x#"$build_dir"}
+	shortname=${x#"$BUILD_DIR"}
 	for i in 1 2; do
 		echo lualatex "$shortname [$i/2]"
 		lualatex --halt-on-error --output-directory out "$x" > /dev/null || exit 13
